Extract ValueType ser/deser and add addValueType() to builder
